Share fuzzing code between the two validation-drt targets
