+++ cedar-drt/fuzz/fuzz_targets/validation-drt.rs
@@ -15,75 +15,8 @@
  */
 
 #![no_main]
-use cedar_drt::*;
-use cedar_drt_inner::*;
-use cedar_policy_core::ast;
-use cedar_policy_generators::{
-    abac::ABACPolicy, hierarchy::HierarchyGenerator, schema::Schema, settings::ABACSettings,
-};
-use libfuzzer_sys::arbitrary::{self, Arbitrary, Unstructured};
-use log::{debug, info};
-
-/// Input expected by this fuzz target
-#[derive(Debug, Clone)]
-pub struct FuzzTargetInput {
-    /// generated schema
-    pub schema: Schema,
-    /// generated policy
-    pub policy: ABACPolicy,
-}
-
-/// settings for this fuzz target
-const SETTINGS: ABACSettings = ABACSettings {
-    match_types: false,
-    enable_extensions: true,
-    max_depth: 7,
-    max_width: 7,
-    enable_additional_attributes: true,
-    enable_like: true,
-    enable_action_groups_and_attrs: true,
-    enable_arbitrary_func_call: true,
-    enable_unknowns: false,
-    enable_action_in_constraints: true,
-};
-
-impl<'a> Arbitrary<'a> for FuzzTargetInput {
-    fn arbitrary(u: &mut Unstructured<'a>) -> arbitrary::Result<Self> {
-        let schema: Schema = Schema::arbitrary(SETTINGS.clone(), u)?;
-        let hierarchy = schema.arbitrary_hierarchy(u)?;
-        let policy = schema.arbitrary_policy(&hierarchy, u)?;
-        Ok(Self { schema, policy })
-    }
-
-    fn try_size_hint(
-        depth: usize,
-    ) -> arbitrary::Result<(usize, Option<usize>), arbitrary::MaxRecursionReached> {
-        Ok(arbitrary::size_hint::and_all(&[
-            Schema::arbitrary_size_hint(depth)?,
-            HierarchyGenerator::size_hint(depth),
-            Schema::arbitrary_policy_size_hint(&SETTINGS, depth),
-        ]))
-    }
-}
+use cedar_drt_inner::validation_drt;
+use libfuzzer_sys::fuzz_target;
 
 // Non-type-directed fuzzing of (strict) validation.
-fuzz_target!(|input: FuzzTargetInput| {
-    initialize_log();
-    let def_impl = LeanDefinitionalEngine::new();
-
-    // generate a schema
-    if let Ok(schema) = ValidatorSchema::try_from(input.schema) {
-        debug!("Schema: {:?}", schema);
-
-        // generate a policy
-        let mut policyset = ast::PolicySet::new();
-        let policy: ast::StaticPolicy = input.policy.into();
-        policyset.add_static(policy).unwrap();
-        debug!("Policies: {policyset}");
-
-        // run the policy through both validators and compare the result
-        let (_, total_dur) =
-            time_function(|| run_val_test(&def_impl, schema, &policyset, ValidationMode::Strict));
-        info!("{}{}", TOTAL_MSG, total_dur.as_nanos());
-    }
-});
+fuzz_target!(|input: validation_drt::FuzzTargetInput<true>| validation_drt::fuzz_target(input));
